Update module files
